+++ include/LIEF/MachO/Binary.hpp
@@ -299,6 +299,13 @@ class LIEF_API Binary : public LIEF::Binary  {
   //! if the section can't be found
   const Section* get_section(const std::string& name) const;
 
+  //! Return the section from the segment with the name
+  //! given in the first parameter and with the section's name provided in the
+  //! second parameter. If the section cannot be found, it returns a nullptr
+  Section* get_section(const std::string& sgname, const std::string& secname);
+
+  const Section* get_section(const std::string& sgname, const std::string& secname) const;
+
   //! Check if a segment with the given name exists
   bool has_segment(const std::string& name) const;